% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAchecks.R
\name{checkJTKperiod}
\alias{checkJTKperiod}
\title{Validate Detection Period for JTK Cycle}
\usage{
checkJTKperiod(time, period)
}
\value{
logical stating whether the period is appropriate for JTK Cycle
for this dataset.
}
\description{
Validate Detection Period for JTK Cycle
}
\keyword{internal}
