% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAchecks.R
\name{checkPeriod}
\alias{checkPeriod}
\title{Validate Detection Period}
\usage{
checkPeriod(time, period)
}
\arguments{
\item{time}{numeric vector of sample collection times.}

\item{period}{hypothesized period.}
}
\value{
logical indicating whether the period is suitable for testing given
the sampling times of the dataset.
}
\description{
Validate Detection Period
}
\keyword{internal}
