% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\name{discoApp}
\alias{discoApp}
\title{Launch the DiscoRhythm Shiny Application}
\usage{
discoApp()
}
\value{
Nothing is returned by this function.
}
\description{
This launches the web interface to DiscoRhythm containing all analysis
tools. The vignette contains details on usage.
}
\examples{
\dontrun{
 discoApp()
}
}
