% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInput.R
\name{discoCheckInput}
\alias{discoCheckInput}
\title{Import Data for DiscoRhythm Analysis}
\usage{
discoCheckInput(Maindata)
}
\arguments{
\item{Maindata}{data.frame with the first column containing row IDs and all
subsequent columns containing experimental values. Columns should follow the
expected naming format described in the vignette. May also be a 
SummarizedExperiment with colData in the same format as Metadata.}
}
\value{
data.frame in Maindata format (checked for errors and
modified as needed)
}
\description{
Performs various checks and cleaning operations on the input data.
}
\details{
Rows containing NA's or all constant values are removed.
If matrix values are character it will be attempted to convert them
to numeric.
If input is not a data.frame it will be converted using
\code{as.data.frame()}.
User will be warned if row IDs contain duplicate entries.
}
\examples{
Maindata <- discoGetSimu()
Maindata_clean <- discoCheckInput(Maindata)

}
