% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{discoColors}
\alias{discoColors}
\title{Color pallete used by DiscoRhythm for plotting.
This palette is duplicated in inst/app/www/custom_styles.css for
application to the shiny app.}
\format{
An object of class \code{list} of length 14.
}
\usage{
discoColors
}
\description{
Color pallete used by DiscoRhythm for plotting.
This palette is duplicated in inst/app/www/custom_styles.css for
application to the shiny app.
}
\keyword{internal}
