% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discoDFtoSE}
\alias{discoDFtoSE}
\alias{discoSEtoDF}
\title{Data formatting for DiscoRhythm}
\usage{
discoDFtoSE(Maindata, Metadata = NULL)

discoSEtoDF(se)
}
\arguments{
\item{Maindata}{data.frame with the first column containing row IDs and all
subsequent columns containing experimental values. Columns should follow the
expected naming format described in the vignette. May also be a 
SummarizedExperiment with colData in the same format as Metadata.}

\item{Metadata}{data.frame of sample data, usually generated by
using \code{discoParseMeta} on the column names of the Maindata data.frame. 
If \code{is.null(Metadata)} and Maindata is provided as input, Metadata will 
be generated from Maindata.}

\item{se}{SummarizedExperiment with \code{!is.null(rownames(se))}.}
}
\value{
a SummarizedExperiment object with colData containing a 
DiscoRhythm format Metadata DataFrame.

a DiscoRhythm format Maindata data.frame.
}
\description{
Functions to convert between usage of
data.frame or SummarizedExperiment data formats.
}
