% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{discoGetSimu}
\alias{discoGetSimu}
\title{Read in the DiscoRhythm Simulated dataset}
\usage{
discoGetSimu()
}
\value{
the simulated demo dataset (Maindata format) used in the DiscoRhythm
web application
}
\description{
A convenience function to get the simulated circadian transcriptomic
system data file used
in DiscoRhythm for various demonstrations and tests.
}
\examples{

indata <- discoGetSimu()

}
