% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discoODAid2name}
\alias{discoODAid2name}
\title{Mapping Identifiers to Full Names}
\format{A named vector, length 4
\describe{
  \item{names(discoODAid2name)}{Identifier}
  \item{as.vector(discoODAid2name)}{Full names}
}}
\usage{
discoODAid2name
}
\description{
A small named vector mapping oscillation detection algorithm names
to a convenient identifier.
}
\keyword{datasets}
