% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{discoPCAgetOutliers}
\alias{discoPCAgetOutliers}
\title{Internal function for applying SD cutoff to PCA results
Returns a logical indicating which samples are not outliers}
\usage{
discoPCAgetOutliers(x, SDfactor = 3, pcToCut = seq_len(4))
}
\value{
logical indicating which samples are outliers in PCA
}
\description{
Internal function for applying SD cutoff to PCA results
Returns a logical indicating which samples are not outliers
}
\keyword{internal}
