% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{discoQC}
\alias{discoQC}
\alias{discoPCAoutliers}
\alias{discoInterCorOutliers}
\alias{discoRepAnalysis}
\title{Quality Control for DiscoRhythm}
\usage{
discoPCAoutliers(Maindata, threshold = 3, scale = TRUE,
  pcToCut = seq_len(4))

discoInterCorOutliers(Maindata, cor_method = "pearson", threshold = 3,
  thresh_type = "SD")

discoRepAnalysis(Maindata, Metadata = NULL,
  aov_method = "Equal Variance", aov_pcut = 0.05, aov_Fcut = 0,
  avg_method = "Median")
}
\arguments{
\item{Maindata}{data.frame with the first column containing row IDs and all
subsequent columns containing experimental values. Columns should follow the
expected naming format described in the vignette. May also be a 
SummarizedExperiment with colData in the same format as Metadata.}

\item{threshold}{numeric, a threshold determining which samples are
outliers (for discoInterCorOutliers, in units of thresh_type, for
discoPCAoutliers in units of standard deviations).}

\item{scale}{logical, perform scaling before PCA (recomended).}

\item{pcToCut}{numeric, which PCs to use for outlier detection. It is
recomended to select the first X PCs based on which PCs explain a
significant amount of variance in the data.}

\item{cor_method}{character, method of pairwise correlation
(see \link[stats]{cor}'s "method" argument for all options).}

\item{thresh_type}{character indicating threshold type (either standard
deviations below the mean, or an absolution correlation value). One of:
"sd" or "value".}

\item{Metadata}{data.frame of sample data, usually generated by
using \code{discoParseMeta} on the column names of the Maindata data.frame. 
If \code{is.null(Metadata)} and Maindata is provided as input, Metadata will 
be generated from Maindata.}

\item{aov_method}{character, method to use for ANOVA. One of:
"Equal Variance", "Welch", or "None".}

\item{aov_pcut}{numeric, p-value cutoff used to select rows with
statistically significant signal-to-noise.}

\item{aov_Fcut}{numeric, F-statistic cutoff used to select rows
with high signal-to-noise based on magnitude.}

\item{avg_method}{character, method for averaging technical replicates. One
of: "Median","Mean","Random", or "None".}
}
\value{
list containing PCA results and the detected outliers

A list of 3 objects:
1) outliers - named boolean indicating if the sample is an outlier
2) meanCor - mean of all pairwise correlations for a given sample
3) corMat - Matrix of all pairwise correlation values
}
\description{
Functions for executing outlier detection and row filtering workflows
prior to rhythmicity analysis.
}
\examples{
Maindata <- discoGetSimu()
pca <- discoPCAoutliers(Maindata)

Maindata <- discoGetSimu()
CorRes <- discoInterCorOutliers(Maindata)

Maindata <- discoGetSimu()
Metadata <- discoParseMeta(colnames(Maindata)[-1])
ANOVAres <- discoRepAnalysis(Maindata,Metadata)

}
