% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{DiscoRhythm-package}
\alias{DiscoRhythm}
\alias{DiscoRhythm-package}
\title{DiscoRhythm: Interactive Workflow for Discovering Rhythmicity in Biological Data}
\description{
Set of functions for estimation of cyclical characteristics, such 
  as period, phase, amplitude, and statistical significance in large temporal 
  datasets. Supporting functions are available for quality control, 
  dimensionality reduction, spectral analysis, and analysis of experimental 
  replicates. Contains a R Shiny web interface to execute all workflow steps.
}
\details{
The main function to run DiscoRhythm in batch mode is [discoBatch()].
Or to access the DiscoRhythm web application use [discoApp()].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewcarlucci/DiscoRhythm}
  \item Report bugs at \url{https://github.com/matthewcarlucci/DiscoRhythm/issues}
}

}
\author{
\strong{Maintainer}: Matthew Carlucci \email{Matthew.Carlucci@camh.ca}

Authors:
\itemize{
  \item Algimantas Kriščiūnas \email{algimantas.krisciunas@gmc.vu.lt}
  \item Haohan Li \email{Xavier.li@edu.uwaterloo.ca}
  \item Povilas Gibas \email{povilas.gibas@bti.vu.lt}
  \item Karolis Koncevičius \email{karolis.koncevicius@gmc.vu.lt}
  \item Art Petronis \email{Art.Petronis@camh.ca}
  \item Gabriel Oh \email{Gabriel.Oh@camh.ca}
}

}
\keyword{internal}
