% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discoShinyHandler}
\alias{discoShinyHandler}
\title{Handle Error/Warning messages appropriately with
shiny notifications for warnings and
pop-ups for errors}
\usage{
discoShinyHandler(expr, section = "Execution", shinySession = NULL)
}
\value{
output from expr
}
\description{
Handle Error/Warning messages appropriately with
shiny notifications for warnings and
pop-ups for errors
}
\keyword{internal}
