% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fisherExact}
\alias{fisherExact}
\title{Extract key values from stats::fisher.test results}
\usage{
fisherExact(var1, var2)
}
\value{
modified output of \link[stats]{fisher.test}
}
\description{
Set p-values of 0 to < 2.2e-16 and reformat odds ratio using formatC
}
\keyword{internal}
