% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAchecks.R
\name{inferFilteredDesign}
\alias{inferFilteredDesign}
\title{DiscoRhythm Experimental Design}
\usage{
inferFilteredDesign(se)
}
\arguments{
\item{se}{SummarizedExperiment, the main data object used by DiscoRhythm 
expected to contain se$ID, se$ReplicateID, se$Time sample metadata and
non-null rownames.}
}
\value{
list with inferred experimental design features needed to perform
replicate analysis and merging in discoRepAnalysis.
}
\description{
Infers the experimental design from various input data
}
\details{
Characteristics of the experiment sampling are gathered to determine which
oscillation deteciton algorithms are suitable.
}
\keyword{internal}
