% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\name{PeriodDetection_range}
\alias{PeriodDetection_range}
\title{Helper for discoPeriodDetection}
\usage{
PeriodDetection_range(times, circular_t, main_per, test_periods)
}
\value{
a set of periods to use for \code{discoPeriodDetection}
}
\description{
Helper for discoPeriodDetection
}
\keyword{internal}
