% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigOperations.R
\name{sincos}
\alias{sincos}
\alias{sincos2acr}
\alias{sincos2amp}
\title{sin/cos cosinor operations}
\usage{
sincos2acr(sin, cos, per = 24)

sincos2amp(sin, cos)
}
\arguments{
\item{sin}{- sine coefficient returned from cosinor model.}

\item{cos}{- cosine coefficient returned from cosinor model.}

\item{per}{- period of oscillation (default = 24).}
}
\value{
acrophase
}
\description{
Converts sine and cosine coeeficients to acrophase/amplitude
}
\examples{
\dontrun{ # don't run since internal
sincos2acr(0.5, 0.5, per = 24)
sincos2amp(0.5, 0.5)
}
}
\author{
Matthew Carlucci
}
\keyword{internal}
