% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\name{discoApp}
\alias{discoApp}
\title{Launch the DiscoRhythm Shiny Application}
\usage{
discoApp(ncores = 1, port = 3838)
}
\arguments{
\item{ncores}{numeric, number of cores to use for parallelized tasks.
Currently, only used in oscillation detection function \code{discoODAs}.}

\item{port}{numeric, port to run the shiny application on. Sets shiny.port
option.}
}
\value{
Nothing is returned by this function.
}
\description{
This launches the web interface to DiscoRhythm containing all analysis
tools. The vignette contains details on usage.
}
\examples{
\dontrun{
 discoApp()
}
}
