% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{discoGetSimu}
\alias{discoGetSimu}
\title{Read in the DiscoRhythm Simulated dataset}
\usage{
discoGetSimu(as_se = FALSE)
}
\arguments{
\item{as_se}{logical, indicates if example data should be returned as a
SummarizedExperiment or data.frame.}
}
\value{
The simulated demo dataset used in the DiscoRhythm
web application as a data.frame or SummarizedExperiment.
}
\description{
A convenience function to get the simulated circadian transcriptomic
system data file used
in DiscoRhythm for various demonstrations and tests.
}
\examples{

indata <- discoGetSimu()

}
