% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAchecks.R, R/disco.R
\name{discoGetODAs}
\alias{discoGetODAs}
\alias{discoODAs}
\alias{discoPeriodDetection}
\title{Get Valid Oscillation Detection Algorithms}
\usage{
discoGetODAs(regressionData, regressionMeta, method = NULL, period,
  circular_t = FALSE)

discoODAs(regressionData, regressionMeta, period = 24, method = "CS",
  circular_t = FALSE, ncores = 1)

discoPeriodDetection(regressionData, regressionMeta, timeType = "linear",
  main_per = 24, test_periods = NULL)
}
\arguments{
\item{regressionData}{data.frame input.}

\item{regressionMeta}{data.frame input.}

\item{method}{character, names of ODAs to use. If length>1 will test all
input method names.}

\item{period}{numeric, the hypothesized period to test for.}

\item{ncores}{number of cores to parallelize with (applicable to JTK, ARSER
and LS only). If 1 will execute in serial.}

\item{timeType}{character, time is either reported as "linear" or
"circular" on some base-cycle (ex. time of day). This determines the periods
that will be tested for.}

\item{main_per}{numeric, if \code{timeType=="circular"} main_per
indicates the period of the base-cycle where sampling times are derived.}

\item{test_periods}{numeric, a vector of the periods to test.
if \code{timeType=="linear"} and length(test_periods)==2 it will be assumed
to be a range of periods to test over.}

\item{realtive_t0}{logical, is time circular on some base-cycle
(ex. time of day).}
}
\value{
A named list of results where each element is a data.frame for the
corresponding method containing estimates for:
p-value, amplitude, acrophase, q-value, and period used.

A data.frame of Rsquared values for each period, for each row of
regressionData.
}
\description{
Runs each selected oscillation deteciton algorithm sequentially to obtain
oscillation characteristics of each row of the input data.
}
\examples{
# Import the simulated example dataset
Maindata <- discoCheckInput(discoGetSimu())
Metadata <- discoParseMeta(colnames(Maindata)[-1])
# Execute the Cosinor method only
discoODAres <- discoODAs(Maindata,Metadata,method="CS")

demodata <- discoGetSimu()
demometa <- discoParseMeta(colnames(demodata)[-1])


# Detect periods on timeType!="circular" data
rsqs <- discoPeriodDetection(demodata,demometa)
# Detect periods on timeType=="circular" data
rsqs <- discoPeriodDetection(demodata,demometa,timeType="circular")

}
\keyword{internal}
