% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAchecks.R, R/disco.R
\name{discoGetODAs}
\alias{discoGetODAs}
\alias{discoODAs}
\title{Get Valid Oscillation Detection Algorithms}
\usage{
discoGetODAs(se, method = NULL, period, circular_t = FALSE)

discoODAs(se, period = 24, method = "CS", circular_t = FALSE,
  ncores = 1)
}
\arguments{
\item{se}{SummarizedExperiment, the main data object used by DiscoRhythm 
expected to contain se$ID, se$ReplicateID, se$Time sample metadata and
non-null rownames. See the vignette for more details.}

\item{method}{character, names of ODAs to use. If length>1 will test all
input method names.}

\item{period}{numeric, the hypothesized period to test for.}

\item{ncores}{number of cores to parallelize with (applicable to JTK, ARSER
and LS only). If 1 will execute in serial.}

\item{realtive_t0}{logical, is time circular on some base-cycle
(ex. time of day).}
}
\value{
A named list of results where each element is a data.frame for the
corresponding method containing estimates for:
p-value, amplitude, acrophase, q-value, and period used.
}
\description{
Runs each selected oscillation deteciton algorithm sequentially to
 obtain
oscillation characteristics of each row of the input data.
 Technical replicates are expected to be merged by this stage.
}
\examples{
# Import the simulated example dataset
se <- discoCheckInput(discoGetSimu(TRUE))
# Execute the Cosinor method only
discoODAres <- discoODAs(se,method="CS")

}
\keyword{internal}
