% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{discoPCA}
\alias{discoPCA}
\title{Perform PCA}
\usage{
discoPCA(Maindata, scale = TRUE, npcs = 10)
}
\arguments{
\item{Maindata}{data.frame with the first column containing row IDs and all
subsequent columns containing experimental values. Columns should follow the
expected naming format described in the vignette.}

\item{scale}{logical, whether or not to scale the data prior to PCA, see
\link[stats]{prcomp} for more details.}

\item{npcs}{numeric, number of principal components to use, see
\link[stats]{prcomp} for more details.}
}
\value{
output from \link[stats]{prcomp} with an added table element
}
\description{
Calculates PCA results from \code{prcomp} with error handling and outputs
suitable for the DiscoRhythm workflow.
}
\examples{
Maindata <- discoGetSimu()
pca <- discoPCA(Maindata[,-1])

}
