% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODAchecks.R
\name{inferFilteredDesign}
\alias{inferFilteredDesign}
\title{DiscoRhythm Experimental Design}
\usage{
inferFilteredDesign(Maindata, Metadata)
}
\arguments{
\item{Maindata}{data.frame with the first column containing row IDs and all
subsequent columns containing experimental values. Columns should follow the
expected naming format described in the vignette.}

\item{Metadata}{data.frame of sample data, usually generated by
using \code{discoParseMeta} on the column names of the Maindata data.frame.}
}
\value{
list with inferred experimental design features needed to perform
replicate analysis and merging in discoRepAnalysis.
}
\description{
Infers the experimental desing from various input data
}
\details{
Characteristics of the experiment sampling are gathered to determine which
oscillation deteciton algorithms are suitable.
}
\keyword{internal}
